styling changes for the components
