+++ nba-fantasy/src/Components/Login-SignUp/Home.css
@@ -1,86 +1,59 @@
-/* Styles for the Home component */
+/* Home.css */
 
-body {
-    background-color: #f5f5f5;
-    font-family: 'Arial', sans-serif;
-    margin: 0;
-    padding: 0;
-  }
-  
-  .container {
-    max-width: 800px;
-    margin: 50px auto;
-    padding: 20px;
-    background-color: #2c3e50;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    color: #ecf0f1;
-  }
-  
-  .header {
-    text-align: center;
-  }
-  
-  .title {
-    font-size: 36px;
-    margin-bottom: 20px;
-    color: #3498db;
-  }
-  
-  .subtitle {
-    font-size: 18px;
-    margin-bottom: 30px;
-    color: #ecf0f1;
-  }
-  
-  .fantasy-buttons {
-    text-align: center;
-  }
-  
-  .fantasy-buttons a {
-    display: inline-block;
-    padding: 15px 30px;
-    background-color: #e74c3c;
-    color: #ecf0f1;
-    text-decoration: none;
-    border-radius: 8px;
-    font-size: 18px;
-    margin: 0 10px;
-    transition: background-color 0.3s;
-  }
-  
-  .fantasy-buttons a:hover {
-    background-color: #c0392b;
-  }
-  
-  .buttons {
-    display: flex;
-    justify-content: center;
-    margin-top: 30px;
-  }
-  
-  .login-button,
-  .signup-button {
-    padding: 15px 30px;
-    margin: 0 20px;
-    text-decoration: none;
-    color: #ecf0f1;
-    border-radius: 8px;
-    font-size: 18px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  
-  .login-button {
-    background-color: #27ae60; /* Green */
-  }
-  
-  .signup-button {
-    background-color: #3498db; /* Blue */
-  }
-  
-  .login-button:hover,
-  .signup-button:hover {
-    background-color: #1d8348; /* Darker Green */
-  }
-  
+
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background: linear-gradient(to top, blanchedalmond, black); /* Semi-transparent white background color */
+  padding: 50px;
+  border-radius: 8px;
+  box-shadow: 0 0 20px rgba(0, 0, 0, 0.2);
+  max-width: 600px;
+  width: 100%;
+  margin: auto; 
+}
+
+.header {
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+.text {
+  font-size: 24px;
+  color: blanchedalmond;
+}
+
+.fantasy-buttons {
+  margin-top: 20px;
+}
+
+.buttons {
+  display: flex;
+  justify-content: space-around; /* Distribute space evenly between buttons */
+  align-items: center;
+  flex-wrap: wrap; /* Allow wrapping to the next line if needed */
+}
+
+.login-button,
+.signup-button {
+  padding: 10px 20px;
+  margin: 10px;
+  text-align: center;
+  text-decoration: none;
+  font-size: 18px;
+  color: #fff; /* White text color */
+  background-color: #ff4500; /* NBA orange */
+  border-radius: 8px;
+  transition: background-color 0.3s ease;
+}
+
+
+
+
+.login-button:hover,
+.signup-button:hover {
+  background-color: blue; /* Slightly darker orange on hover */
+}
+
