+++ nba-fantasy/src/Components/Sports/Player.css
@@ -1,0 +1,78 @@
+/* Players.css */
+
+.container {
+  max-width: 800px;
+  margin: 0 auto;
+  padding: 20px;
+  background: #f0f0f0;
+}
+  
+  .header {
+    text-align: center;
+    margin-bottom: 20px;
+    color: #007bff; /* NBA blue */
+  }
+  
+  .player-list {
+    list-style-type: none;
+    padding: 0;
+  }
+  
+  .player-list li {
+    margin-bottom: 20px;
+    background-color: #fff;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    border-radius: 8px;
+    overflow: hidden;
+  }
+  
+  .player-link {
+    display: block;
+    padding: 20px;
+    text-decoration: none;
+    color: #333;
+    transition: transform 0.3s ease;
+  }
+  
+  .player-link:hover {
+    transform: scale(1.05);
+  }
+  
+  .search-bar {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 20px;
+  }
+  
+  .search-input {
+    padding: 15px;
+    font-size: 18px;
+    border: 2px solid #007bff; /* NBA blue */
+    border-radius: 8px;
+    margin-right: 10px;
+  }
+  
+  .search-button {
+    padding: 15px;
+    background-color: #007bff; /* NBA blue */
+    color: #fff;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+    font-size: 18px;
+    transition: background-color 0.3s ease;
+  }
+  
+  .search-button:hover {
+    background-color: black; /* Slightly darker blue */
+  }
+  
+  .loading {
+    text-align: center;
+    margin-top: 20px;
+    font-size: 20px;
+    color: #007bff; /* NBA blue */
+  }
+  
+
