+++ nba-fantasy/src/Components/Sports/Team.css
@@ -1,0 +1,105 @@
+/* Team.css */
+
+.TeamWrapper {
+    display: flex;
+    max-width: 800px;
+    margin: 0 auto;
+    padding: 20px;
+  }
+  
+  .Team {
+    flex: 1;
+  }
+  
+  .Scoreboard {
+    flex: 0 0 200px; 
+    margin-left: 20px;
+    padding: 20px;
+    background-color: #1b1b1b; 
+    color: #fff; 
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: column;
+    align-items: center; 
+    justify-content: center; 
+  }
+  
+  .Scoreboard h2 {
+    font-size: 18px;
+    margin-bottom: 10px;
+  }
+  
+  .Scoreboard p {
+    font-size: 24px;
+    font-weight: bold;
+    margin: 0;
+  }
+  .Team h1 {
+    text-align: center;
+    color: #ff4500; /* NBA orange */
+    margin-bottom: 20px;
+  }
+  
+  .todo-input {
+    width: 70%;
+    padding: 10px;
+    margin-right: 10px;
+    font-size: 16px;
+    border: 2px solid #007bff; /* NBA blue */
+    border-radius: 8px;
+    color: #333;
+  }
+  
+  .todo-btn {
+    padding: 10px 15px;
+    background-color: #007bff; /* NBA blue */
+    color: #fff;
+    font-size: 16px;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+  }
+  
+  .todo-btn:hover {
+    background-color: #0056b3; /* Slightly darker blue on hover */
+  }
+  
+  .alert {
+    background-color: #ff4500; /* NBA orange */
+    color: #fff;
+    padding: 10px;
+    margin-top: 10px;
+    text-align: center;
+    border-radius: 8px;
+  }
+  
+  .TeamDisplay {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: #fff;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    border-radius: 8px;
+    padding: 10px;
+    margin-top: 10px;
+  }
+  
+  .TeamDisplay p {
+    margin: 0;
+    font-size: 16px;
+  }
+  
+  .TeamDisplay div {
+    display: flex;
+    gap: 10px;
+    cursor: pointer;
+  }
+  
+ 
+
+
+  
+
+  
+
