+++ nba-fantasy/src/styles.css
@@ -1,0 +1,8 @@
+
+
+
+body {
+  margin: 0;
+  font-family: "Harlow Solid Italic", sans-serif;
+  background: linear-gradient(to left, white, orangered);
+}
